+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bump dependencies
 - Allow sprite frames to have different sizes
 - Fix missing color in index-alpha sprite extract
+- Add --no-alpha flag to sprite extract
 
 # v1.5.0
 - Add "map neat" entity preprocessor command